requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
pyjwt
pyotp
jinja2
python-dotenv